update contract  test/Vipsland.js

+++ contracts/Vipsland.sol
@@ -146,12 +146,12 @@ contract Vipsland is ERC1155Supply, Ownable, ReentrancyGuard, PaymentSplitter {
     uint[] public intArrPRTAIRDROP;
 
     //toggle start
-    uint public presalePRT = 0;
+    uint8 public presalePRT = 0;
     bool public mintMPIsOpen = false;
     bool public mintInternalTeamMPIsOpen = false;
     bool public mintAirdropMPIsOpen = false;
 
-    function setPreSalePRT(uint num) public onlyOwner onlyAllowedNum(num) {
+    function setPreSalePRT(uint8 num) public onlyOwner onlyAllowedNum(num) {
         presalePRT = num;
     }
